cambios xd

+++ InmoSoft/appInmoSoft/models.py
@@ -172,9 +172,9 @@ class PlanDePago(models.Model):
     
 class RegistroPago(models.Model):
     regFechaPago=models.DateField(auto_now=True, db_comment="Fecha de pago")
-    regValorPago=models.IntegerField(db_comment="Valor de pagode la cuota")
-    regPendiente=models.IntegerField(db_comment="Valor pendiente")
-    regRecaudo= models.IntegerField(db_comment="Recaudo total")
+    regValorPago=models.BigIntegerField(db_comment="Valor de pagode la cuota")
+    regPendiente=models.BigIntegerField(db_comment="Valor pendiente")
+    regRecaudo= models.BigIntegerField(db_comment="Recaudo total")
     regfechaHoraCreacion  = models.DateTimeField(auto_now_add=True,db_comment="Fecha y hora del registro")
     regfechaHoraActualizacion = models.DateTimeField(auto_now=True,db_comment="Fecha y hora última actualización")
     regPlanDePago = models.ForeignKey(PlanDePago, on_delete=models.PROTECT, db_comment="plan de pago")
